requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.0
Pillow==5.4.1
scikit-learn==0.20.2
scipy==1.2.0
sklearn==0.0
Werkzeug==0.14.1
